+++ CHANGELOG.md
@@ -10,9 +10,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- id parameter for command /check-quotes
+
 ### Changed
 
+- better random for quote selection
+
 ### Fixed
+
+- Removed check-quotes select menu as the maximum size is 25
 
 ## [2.1.1]